base

+++ src/Class10.java
@@ -12,6 +12,6 @@ public final class Class10
     int anInt277;
     public Class30_Sub2_Sub4 aClass30_Sub2_Sub4_278;
     public Class30_Sub2_Sub4 aClass30_Sub2_Sub4_279;
-    public int anInt280;
+    public long anInt280;
     byte aByte281;
 }
