+++ src/Class30_Sub2_Sub4.java
@@ -4,7 +4,7 @@ public class Class30_Sub2_Sub4 extends Class30_Sub2
 {
 
     public void method443(int i, int j, int k, int l, int i1, int j1, int k1, 
-            int l1, int i2)
+            int l1, long i2)
     {
         Class30_Sub2_Sub4_Sub6 class30_sub2_sub4_sub6 = method444(4016);
         if(class30_sub2_sub4_sub6 != null)
