+++ src/Class49.java
@@ -9,6 +9,6 @@ public final class Class49
     int anInt812;
     int anInt813;
     public Class30_Sub2_Sub4 aClass30_Sub2_Sub4_814;
-    public int anInt815;
+    public long anInt815;
     byte aByte816;
 }
